refactor header structure
