--- src/components/logo/logo-link.tsx
@@ -1,19 +1,0 @@
-import { Link } from 'react-router-dom';
-import { AppRoute } from '../../utils/data';
-
-export default function LogoLink() {
-  return (
-    <Link
-      className="header__logo-link"
-      to={AppRoute.Root}
-    >
-      <img
-        className="header__logo"
-        src="img/logo.svg"
-        alt="6 cities logo"
-        width="81"
-        height="41"
-      />
-    </Link>
-  );
-}
