--- src/components/logo/logo.tsx
@@ -1,13 +1,0 @@
-export default function Logo() {
-  return (
-    <a className="header__logo-link header__logo-link--active">
-      <img
-        className="header__logo"
-        src="img/logo.svg"
-        alt="6 cities logo"
-        width="81"
-        height="41"
-      />
-    </a>
-  );
-}
